feature: restructured sub commands to canonical structure
