--- docs/stackit_beta_security-group_group.md
@@ -1,37 +1,0 @@
-## stackit beta security-group group
-
-Manage security groups
-
-### Synopsis
-
-Manage the lifecycle of security groups.
-
-```
-stackit beta security-group group [flags]
-```
-
-### Options
-
-```
-  -h, --help   Help for "stackit beta security-group group"
-```
-
-### Options inherited from parent commands
-
-```
-  -y, --assume-yes             If set, skips all confirmation prompts
-      --async                  If set, runs the command asynchronously
-  -o, --output-format string   Output format, one of ["json" "pretty" "none" "yaml"]
-  -p, --project-id string      Project ID
-      --verbosity string       Verbosity of the CLI, one of ["debug" "info" "warning" "error"] (default "info")
-```
-
-### SEE ALSO
-
-* [stackit beta security-group](./stackit_beta_security-group.md)	 - Manage security groups
-* [stackit beta security-group group create](./stackit_beta_security-group_group_create.md)	 - Creates security groups
-* [stackit beta security-group group delete](./stackit_beta_security-group_group_delete.md)	 - Deletes a security group
-* [stackit beta security-group group describe](./stackit_beta_security-group_group_describe.md)	 - Describes security groups
-* [stackit beta security-group group list](./stackit_beta_security-group_group_list.md)	 - Lists security groups
-* [stackit beta security-group group update](./stackit_beta_security-group_group_update.md)	 - Updates a security group
-
